Changed data type to decimal and reduced amount of retries to 11

+++ DeviceSocket.cs
@@ -18,7 +18,7 @@ namespace ClimateMeter.Device.Net
             return Task.FromResult(Guid.Empty);
         }
 
-        public Task AddSensorReading(Guid deviceId, float temperature, float humidity)
+        public Task AddSensorReading(Guid deviceId, decimal temperature, decimal humidity)
         {
             return Task.CompletedTask;
         }
